make it compatible with rspec version >= 2.12
fix the WARNING: let declaration xxx accessed in a `before(:all)` hook at:

This is deprecated behavior that will not be supported in RSpec 3.

`let` and `subject` declarations are not intended to be called
in a `before(:all)` hook, as they exist to define state that
is reset between each example, while `before(:all)` exists to
define state that is shared across examples in an example group.